add np.float

+++ test.py
@@ -5,6 +5,5 @@ import numpy as np
 with open('model.pkl', 'rb') as f:
     model = pickle.load(f)
 
-# Predict
-sample = np.array([[0.5, -0.2, 0.1, 0.8]])
+sample = np.array([[0.5, -0.2, 0.1, 0.8]], dtype=np.float)
 print(f"Prediction: {model.predict(sample)}")
